backend/requirements.txt
asgiref==3.8.1
Django==5.1.4
sqlparse==0.5.2
django-cors-headers==4.4.0
django-extensions==3.2.3
django-redis==5.4.0
django-safedelete==1.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
ipython==8.26.0
python-dotenv==1.0.1
requests==2.32.3
